Create world ai file
