+++ src/game/worlds/gameWorldFactory.ts
@@ -15,34 +15,16 @@ import { Map, MapBuilder } from '@game/map/mapBuilder'
 import AISystem from '@game/systems/aiSystem'
 import InvincibleSystem from '@game/systems/invincibleSystem'
 import { DamageSystem } from '@game/systems/damageSystem'
-import { FamilyBuilder } from '@core/ecs/family'
 import { AirHolderSystem } from '@game/systems/airHolderSystem'
 import * as PIXI from 'pixi.js'
-import { TitleWorldFactory } from './titleWorldFactory'
-import { isAlive } from '@game/ai/entity/common/condition/isAlive'
-import { assert } from '@utils/assertion'
-import { wait } from '@core/behaviour/wait'
 import { FilterSystem } from '@game/systems/filterSystem'
 import { LightSystem } from '@game/systems/lightSystem'
 import { EventSensorSystem } from '@game/systems/eventSensorSystem'
-import { Behaviour } from '@core/behaviour/behaviour'
-
-const gameWorldBehaviour = function*(world: World): Behaviour<World> {
-  const playerFamily = new FamilyBuilder(world).include('Player').build()
-  assert(playerFamily.entityArray.length === 1)
-  const playerEntity = playerFamily.entityArray[0]
-  const isPlayerAlive = isAlive(playerEntity)
-
-  while (isPlayerAlive()) {
-    yield
-  }
-  yield* wait(60)
-  return new TitleWorldFactory().create()
-}
+import { gameWorldAI } from '@game/ai/world/game/gameWorldAI'
 
 export class GameWorldFactory {
   public create(map: Map, playerSpawnerID: number): World {
-    const world = new World(gameWorldBehaviour)
+    const world = new World(gameWorldAI)
 
     const gameWorldContainer = new Container()
     world.stage.addChild(gameWorldContainer)
